polki_dolki/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==1.0.0
Django==4.1.2
django-filter==22.1
django-rest-swagger==2.2.0
django-swagger-ui==0.1.11
djangorestframework==3.14.0
idna==3.4
importlib-metadata==5.0.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
openapi-codec==1.3.2
psycopg2==2.9.4
Pygments==2.13.0
pytz==2022.4
PyYAML==6.0
requests==2.28.1
signals==0.0.2
simplejson==3.17.6
sqlparse==0.4.3
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.12
zipp==3.10.0
